+++ azure-storage-blob/ChangeLog.md
@@ -2,6 +2,10 @@
 
 > See [BreakingChanges](BreakingChanges.md) for a detailed list of API breaks.
 
+## Version XX.XX.XX:
+
+- Optimized page blob upload for create_blob_from_* methods, by skipping the empty chunks.
+
 ## Version 0.37.1:
 
 - Enabling MD5 validation no longer uses the memory-efficient algorithm for large block blobs, since computing the MD5 hash requires reading the entire block into memory.